Fully Document the Events Frontend Feature
This commit fully documents the events frontend feature, including components, widgets, services, etc.
