+++ frontend/src/app/event/event-filter/event-filter.pipe.ts
@@ -7,7 +7,7 @@
  */
 
 import { Pipe, PipeTransform } from '@angular/core';
-import { Event } from '../event.service';
+import { Event } from '../event.model';
 
 @Pipe({
   name: 'eventFilter'
