+++ frontend/src/app/event/event-page/event-page.component.html
@@ -1,5 +1,4 @@
 <div>
-
     <div class="parameter-bar">
         <search-bar class="search-bar" [searchBarQuery]="searchBarQuery"
             (searchBarQueryChange)="onSearchBarQueryChange($event)" />
