+++ frontend/src/app/event/event-routing.module.ts
@@ -1,3 +1,12 @@
+/**
+ * The Event Routing Module holds all of the routes that are children
+ * to the path /events/...
+ * 
+ * @author Ajay Gandecha, Jade Keegan, Brianna Ta, Audrey Toney
+ * @copyright 2023
+ * @license MIT
+ */
+
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { EventDetailsComponent } from './event-details/event-details.component';
