+++ frontend/src/app/event/event.service.ts
@@ -10,25 +10,13 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Organization } from '../organization/organization.service';
-
-/** Interface for Event Type (used on frontend for event detail) */
-export interface Event {
-    id: number | null;
-    name: string;
-    time: Date;
-    location: string;
-    description: string;
-    public: boolean;
-    organization_id: number;
-    organization: Organization;
-}
+import { Event } from './event.model';
 
 @Injectable({
     providedIn: 'root'
 })
 export class EventService {
-    
+  
   constructor(protected http: HttpClient) {}
 
   /** Returns all event entries from the backend database table using the backend HTTP get request.
