+++ frontend/src/app/event/widgets/event-card/event-card.widget.css
@@ -2,7 +2,6 @@
 .mat-mdc-card {
     margin: 0 !important;
     max-width: 100%;
-    /* background-color: #525252 !important; */
 }
 
 .mat-mdc-card:hover {
