+++ frontend/src/app/event/widgets/event-card/event-card.widget.html
@@ -1,6 +1,8 @@
+<!-- Event Card -->
 <mat-card [ngClass]="(selected && disableLink) ? 'selected-card' : ''" appearance="outlined"
     style="margin-top: 12px !important;" [routerLink]="disableLink ? null : '/events/' + event.id"
     (click)="cardClicked()">
+    <!-- Inner Columns -->
     <div style="display: flex; flex-direction: column; padding: 12px;">
         <div style="display: flex; flex-direction: row;">
             <p style="font-weight: 500; margin: 0px; width: 75%;">{{event.name}}</p>
@@ -10,6 +12,5 @@
             <p style="margin: 0px; width: 50%;">{{event.organization.shorthand}}</p>
             <p style="margin: 0px; text-align: end; width: 50%;">{{event.location}}</p>
         </div>
-        <!-- <p class="description-text">{{event.description}}</p> -->
     </div>
 </mat-card>
