+++ frontend/src/app/event/widgets/event-detail-card/event-detail-card.widget.ts
@@ -1,5 +1,5 @@
 /**
- * The Event Card widget abstracts the implementation of the
+ * The Event Detail Card widget abstracts the implementation of the
  * detail event card from the whole event page.
  * 
  * @author Ajay Gandecha
@@ -8,8 +8,7 @@
  */
 
 import { Component, Input } from '@angular/core';
-import { Organization } from 'src/app/organization/organization.service';
-import { Event } from '../../event.service';
+import { Event } from '../../event.model';
 import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
@@ -19,12 +18,20 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 })
 export class EventDetailCard {
 
+    /** The event for the event card to display */
     @Input() event!: Event
 
+    /** Constructs the widget */
     constructor(protected snackBar: MatSnackBar) { }
 
+    /** Handler for when the share button is pressed
+     *  This function copies the permalink to the event to the user's
+     *  clipboard.
+     */
     onShareButtonClick() {
+        // Write the URL to the clipboard
         navigator.clipboard.writeText("https://csxl.unc.edu/events/" + this.event.id);
+        // Open a snackbar to alert the user
         this.snackBar.open("Event link copied to clipboard.", "", {duration: 3000})
     }
 }
